requirements.txt
python-telegram-bot==13.7
matplotlib==3.5.1
numpy==1.22.2
python-dotenv==0.19.2